refactor: extract LogPresentation+StepLogActionButtonStyle to own file, fix file_length (#2911)
